+++ src/main/java/com/example/spring_jscode/domain/repository/BoardRepository.java
@@ -1,6 +1,6 @@
-package com.example.spring_jscode.repository;
+package com.example.spring_jscode.domain.repository;
 
-import com.example.spring_jscode.entity.Board;
+import com.example.spring_jscode.domain.entity.Board;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
