+++ src/main/java/com/example/spring_jscode/dto/BoardPageResponseDto.java
@@ -3,17 +3,18 @@ package com.example.spring_jscode.dto;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
+import org.springframework.data.domain.Page;
 
 import java.util.List;
 
 @Getter
 @AllArgsConstructor(access = AccessLevel.PRIVATE)
 public class BoardPageResponseDto {
-    private List<BoardResponseDto> boardResponseDtos;
+    private Page<BoardResponseDto> boardResponseDtos;
     private Integer totalPages;
     private Integer offset;
 
-    public static BoardPageResponseDto toDtoFromBoardResponseDto(List<BoardResponseDto> boardResponseDtos, Integer totalPages, Integer offset) {
+    public static BoardPageResponseDto toDtoFromBoardResponseDto(Page<BoardResponseDto> boardResponseDtos, Integer totalPages, Integer offset) {
         return new BoardPageResponseDto(boardResponseDtos, totalPages, offset);
     }
 }
