+++ src/main/java/com/example/spring_jscode/dto/BoardRequestDto.java
@@ -1,11 +1,14 @@
 package com.example.spring_jscode.dto;
 
+import com.example.spring_jscode.domain.entity.Board;
 import lombok.Getter;
-import lombok.Setter;
 
 @Getter
-@Setter
 public class BoardRequestDto {
     private String title;
     private String content;
+
+    public Board toEntity(BoardRequestDto boardRequestDto) {
+        return new Board(boardRequestDto.getTitle(), boardRequestDto.getContent());
+    }
 }
